+++ cypress/integration/elements-selection.spec.js
@@ -3,15 +3,15 @@ import authValid from '../fixtures/authValid'
 import emptyNotes from '../fixtures/emptyNotes'
 import graphResponse from '../fixtures/graphResponse'
 import getStyling from '../fixtures/getStyling'
-import { ROUTE_NODES_SELECTION } from '../../src/constants/routes'
+import {ROUTE_ELEMENTS_SELECTION} from '../../src/constants/routes'
 
-context('Nodes selection', () => {
+context('Elements selection', () => {
   beforeEach(() => {
     cy.visit('http://localhost:3000')
   })
 
-  describe('Nodes selection', () => {
-    it('Nodes selection should return results', () => {
+  describe('Elements selection', () => {
+    it('Elements selection should return results', () => {
       cy.intercept({
         method: 'POST',
         url: '**/login',
@@ -29,7 +29,7 @@ context('Nodes selection', () => {
 
       cy.intercept({
         method: 'GET',
-        url: '**/graph/edges/notes',
+        url: '**/graph/elements/notes',
       }, emptyNotes).as('getEdgesNotes')
 
       cy.intercept({
@@ -69,27 +69,49 @@ context('Nodes selection', () => {
       cy.get('.nav-left').should('contain', 'Edges: 14')
 
       // click the nodes selection icon
-      cy.get('#sidebar-button-nodes-selection').click()
+      cy.get('#sidebar-button-elements-selection').click()
 
-      cy.location('pathname').should('be.equal', ROUTE_NODES_SELECTION)
+      cy.wait(500)
 
-      cy.get('.nodes-selection-details').should('not.have.exist')
+      cy.location('pathname').should('be.equal', ROUTE_ELEMENTS_SELECTION)
+
+      cy.get('.elements-selection-details').should('not.have.exist')
 
       // select first node
-      cy.get('#node-select').find('.p-dropdown-trigger').click({ force: true })
-      cy.get('#node-select').find('.p-dropdown-filter').type('asset')
-      cy.get('#node-select').find('.p-dropdown-item').eq(0).click({ force: true })
+      cy.get('#element-select').find('.p-dropdown-trigger').click({ force: true })
+      cy.get('#element-select').find('.p-dropdown-filter').type('asset')
+      cy.get('#element-select').find('.p-dropdown-item').eq(0).click({ force: true })
 
       cy.get('.nodes-selection-details-table-properties').find('tbody tr').should('have.length', '7')
       cy.get('.nodes-selection-details-table-relationships').find('tbody tr').should('have.length', '10')
 
       // select another node
-      cy.get('#node-select').find('.p-dropdown-trigger').click({ force: true })
-      cy.get('#node-select').find('.p-dropdown-filter').clear().type('str')
-      cy.get('#node-select').find('.p-dropdown-item').eq(0).click({ force: true })
+      cy.get('#element-select').find('.p-dropdown-trigger').click({ force: true })
+      cy.get('#element-select').find('.p-dropdown-filter').clear().type('str')
+      cy.get('#element-select').find('.p-dropdown-item').eq(0).click({ force: true })
 
       cy.get('.nodes-selection-details-table-properties').find('tbody tr').should('have.length', '8')
       cy.get('.nodes-selection-details-table-relationships').find('tbody tr').should('have.length', '2')
+
+      // change type to edge
+      cy.get('#element-type-select').find('.p-button').eq(1).click()
+
+      // select first edge
+      cy.get('#element-select').find('.p-dropdown-trigger').click({ force: true })
+      cy.get('#element-select').find('.p-dropdown-filter').clear().type('asset')
+      cy.get('#element-select').find('.p-dropdown-item').eq(0).click({ force: true })
+
+      cy.get('.edges-selection-details-table-properties').find('tbody tr').should('have.length', '2')
+      cy.get('.edges-selection-details-table-relationships').find('tbody tr').should('have.length', '1')
+
+      // select another edge
+      cy.get('#element-select').find('.p-dropdown-trigger').click({ force: true })
+      cy.get('#element-select').find('.p-dropdown-filter').clear().type('str')
+      cy.get('#element-select').find('.p-dropdown-item').eq(0).click({ force: true })
+
+      cy.get('.edges-selection-details-table-properties').find('tbody tr').should('have.length', '2')
+      cy.get('.edges-selection-details-table-relationships').find('tbody tr').should('have.length', '1')
+
     })
   })
 })